requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.1
async-timeout==5.0.1
attmap==0.13.2
attrs==25.4.0
av==16.0.1
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.1
decorator==5.2.1
distro==1.9.0
eido==0.2.4
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.122.0
faster-whisper==1.2.1
filelock==3.20.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
genai==2.1.0
google-auth==2.43.0
google-genai==1.52.0
groq==0.36.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.5
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.11
intel-openmp==2021.4.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
logmuse==0.2.8
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.4.2
numpy==2.2.6
onnxruntime==1.23.2
openai==0.27.10
opencv-python==4.12.0.88
oyaml==1.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pephubclient==0.4.5
peppy==0.40.8
pillow==12.0.0
piper==0.14.5
pipestat==0.12.2
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.29.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
scipy==1.15.3
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
speechbrain==1.0.3
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tabulate==0.9.0
tbb==2021.13.1
tenacity==9.1.2
tiktoken==0.3.3
tokenizers==0.22.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ubiquerg==0.8.1
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
webrtcvad-wheels==2.0.14
websockets==15.0.1
yacman==0.9.4
yarl==1.22.0
